+++ module/Application/src/Application/Form/DiffusionTheseForm.php
@@ -5,8 +5,8 @@ namespace Application\Form;
 use Application\Entity\Db\Diffusion;
 use Application\Entity\Db\NatureFichier;
 use Application\Entity\Db\VersionFichier;
-use Application\Service\Fichier\FichierServiceAwareInterface;
-use Application\Service\Fichier\FichierServiceAwareTrait;
+use Application\Service\FichierThese\FichierTheseServiceAwareInterface;
+use Application\Service\FichierThese\FichierTheseServiceAwareTrait;
 use Application\Service\Message\DiffusionMessages;
 use UnicaenApp\Exception\LogicException;
 use UnicaenApp\Form\Element\Date;
@@ -18,10 +18,10 @@ use Zend\InputFilter\InputFilterProviderInterface;
 use Zend\Validator\Callback;
 
 class DiffusionTheseForm extends Form
-    implements InputFilterProviderInterface, MessageServiceAwareInterface, FichierServiceAwareInterface
+    implements InputFilterProviderInterface, MessageServiceAwareInterface, FichierTheseServiceAwareInterface
 {
     use MessageServiceAwareTrait;
-    use FichierServiceAwareTrait;
+    use FichierTheseServiceAwareTrait;
 
     /**
      * @var VersionFichier
@@ -362,8 +362,7 @@ class DiffusionTheseForm extends Form
                                 /** @var Diffusion $diffusion */
                                 $diffusion = $this->getObject();
                                 $besoinVersionExpurgee = $value === Diffusion::DROIT_AUTEUR_OK_NON;
-//                                $theseExpurgeeDeposee = $diffusion->getThese()->getFichiersBy(false, true, false, $this->versionFichier)->count() > 0;
-                                $theseExpurgeeDeposee = ! empty($this->fichierService->getRepository()->fetchFichiers($diffusion->getThese(), NatureFichier::CODE_THESE_PDF, $this->versionFichier, false));
+                                $theseExpurgeeDeposee = ! empty($this->fichierTheseService->getRepository()->fetchFichierTheses($diffusion->getThese(), NatureFichier::CODE_THESE_PDF, $this->versionFichier, false));
 
                                 // des fichiers expurgés doivent avoir été déposés en cas de pb de droit d'auteur
                                 if ($besoinVersionExpurgee && !$theseExpurgeeDeposee) {