+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 Journal des modifications
 =========================
 
+1.2.0 (10/07/2019)
+------------------
+
+### Améliorations
+
+- Amélioration des temps de réponse du moteur de workflow en base de données.
+- Refonte technique de la gestion des fichiers liés aux thèses, prélable au travail sur les droits de dépôt de 
+  fichiers divers.
+
+
 1.1.9 (14/05/2019)
 ------------------